refactor TestOutput into stream
