+++ cli/alsatian-cli.ts
@@ -1,12 +1,12 @@
 #! /usr/bin/env node
 
-import { createTestSet, TestRunner, TestOutput } from "../core/alsatian-core";
+import { createTestSet, TestRunner, TestOutputStream } from "../core/alsatian-core";
 import { CliTestRunner } from "./cli-test-runner";
 import { AlsatianCliOptions } from "./alsatian-cli-options";
 import { Writable, Readable } from "stream";
 
 const {
-  TapBark
+   TapBark
 } = require("tap-bark");
 
 // get all arguments from the user
@@ -16,23 +16,22 @@ let userArguments = new AlsatianCliOptions(process.argv.slice(2));
 let testSet = createTestSet();
 testSet.addTestsFromFiles(userArguments.fileGlobs);
 
-const stream = new Readable();
+let outputStream = new TestOutputStream();
 
 if (userArguments.tap) {
-  // if they want TAP output then just write to stdout directly
-  stream.pipe(process.stdout);
-} else {
-  // otherwise create the tap bark reporter
-  const bark = TapBark.create();
+   // if they want TAP output then just write to stdout directly
+   outputStream.pipe(process.stdout);
+}
+else {
+   // otherwise create the tap bark reporter
+   const bark = TapBark.create();
 
-  // pipe the reporter into stdout
-  stream.pipe(bark.getPipeable()).pipe(process.stdout);
+   // pipe the reporter into stdout
+   outputStream.pipe(bark.getPipeable()).pipe(process.stdout);
 }
 
-let output = new TestOutput(stream);
-
 // create alsatian test runner
-let testRunner = new TestRunner(output);
+let testRunner = new TestRunner(outputStream);
 
 // run the test set
 let cliTestRunner = new CliTestRunner(testRunner);
