+++ core/_core.ts
@@ -5,7 +5,7 @@ import { createTestSet } from "./create-test-set";
 import { FileRequirer } from "./file-requirer";
 import { GlobHelper } from "./glob-helper";
 import { TestFixture } from "./test-fixture";
-import { TestOutput } from "./test-output";
+import { TestOutputStream } from "./test-output-stream";
 
 export {
     createTestSet,
@@ -15,5 +15,5 @@ export {
     FileRequirer,
     GlobHelper,
     TestFixture,
-    TestOutput
+    TestOutputStream
 };
